# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DVinyl API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Mobile/native client API for a DVinyl instance, covering authentication,
    collection, wishlist and sharing management, the full item lifecycle, plugin
    metadata, instance administration, and account settings.

    All routes are mounted under `/api/v1`. Requests and responses are JSON
    (`application/json`), except the file-upload endpoints (`multipart/form-data`),
    the share-link QR endpoint (`image/png`) and the import/export endpoints, whose
    exports answer CSV, ZIP or JSON file downloads.

    Authentication is a JWT Bearer access token obtained from `POST /auth/login`.
    Access tokens live 15 minutes (`expiresIn: 900`); use the rotating refresh
    token from `POST /auth/refresh`. Collection-scoped routes are additionally
    gated on the caller's role in the collection named in the path
    (`viewer` < `editor` < `admin`); instance admins pass every collection role.
    Admin routes require `isAdmin` on the account.

    Error envelope: every failed response is `{ "success": false, "error": "..." }`
    with the HTTP status appropriate to the cause. Success responses are
    resource-shaped (e.g. `{ "item": ... }`, `{ "items": [...], "page": ... }`).

    On an instance that has not completed first-run setup, every `/api/v1`
    request returns `503` with `{ "success": false, "error": "Instance not set up yet" }`.
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /collections/{id}/lists/{listId}/candidates:
    get:
      tags: [Lists]
      operationId: listCandidates
      summary: Search items or tracks to add to a list
      description: |
        Requires at least `editor` role. For an items list, matches title, barcode
        and each enabled plugin's creator field; for a playlist, matches track
        titles and returns one row per track. Results carry `inList`. An empty or
        absent `q` returns no results.
      parameters:
        - $ref: '#/components/parameters/CollectionId'
        - $ref: '#/components/parameters/ListId'
        - name: q
          in: query
          required: false
          description: Search text (trimmed, first 100 chars).
          schema: { type: string }
      responses:
        '200':
          description: Candidates.
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/ListCandidate' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: No such list in this collection.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { success: false, error: 'List not found' }
        '500': { $ref: '#/components/responses/InternalServerError' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `POST /auth/login` (15-minute lifetime).
  parameters:
    CollectionId:
      name: id
      in: path
      required: true
      description: Collection id (Mongo ObjectId).
      schema: { type: string }
...
    ListId:
      name: listId
      in: path
      required: true
      description: List id (Mongo ObjectId).
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: Missing, invalid or expired Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing:
              value: { success: false, error: 'Missing bearer token' }
            invalid:
              value: { success: false, error: 'Invalid or expired token' }
    Forbidden:
      description: Authenticated, but lacking the required role or instance-admin flag.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Forbidden' }
...
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example: { success: false, error: 'Internal server error' }
...
  schemas:
    Error:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: string }
        code:
          type: string
          description: Optional machine-readable code; not used by every handler.
      example: { success: false, error: 'Invalid email or password' }
...
    ListCandidate:
      type: object
      additionalProperties: true
      properties:
        item: { type: string }
        title: { type: string }
        creator: { type: string }
        details: { type: string }
        cover: { type: string }
        inList: { type: boolean }
        track: { type: string }
        trackTitle: { type: string }
        position: { type: string }